+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog][keepachangelog] and this project adheres to [Semantic Versioning][semver].
 
+## UNRELEASED
+
+### Added
+
+- Listener `FlushTranslatorCacheListener` for memory leak fixing on `Translator` implementation [#69]
+
+[#69]:https://github.com/spiral/roadrunner-laravel/pull/69
+
 ## v5.4.0
 
 ### Added