+++ CHANGELOG.md
@@ -4,13 +4,20 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog][keepachangelog] and this project adheres to [Semantic Versioning][semver].
 
+## UNRELEASED
+
+### Added
+
+- Listener `CleanupUploadedFilesListener` for removing temporary files, which was created during uploading _(should be enabled manually for the `AfterLoopIterationEvent` event)_ [#84]
+
+[#84]:https://github.com/spiral/roadrunner-laravel/issues/84
+
 ## v5.7.0
 
 ### Added
 
 - Laravel 9 support [#78]
 - Listener `FlushStrCacheListener` for flushing `Str` cache between requests [#86]
-- Listener `CleanupUploadedFilesListener` for removing temporary files, which was created during uploading [#84]
 
 ### Removed
 
@@ -18,7 +25,6 @@ The format is based on [Keep a Changelog][keepachangelog] and this project adher
 
 [#78]:https://github.com/spiral/roadrunner-laravel/pull/78
 [#86]:https://github.com/spiral/roadrunner-laravel/pull/86
-[#84]:https://github.com/spiral/roadrunner-laravel/issues/84
 
 ## v5.6.0